fixed bugs in the spreadsheet and CSV file
Specifically:

hash_analysis.xlsx -- several inter-sheet references used a . separator,
e.g., data.A1, which is not supported by Excel but works in Libre
Office. I changed these to all use !, e.g., data!A1. That seemed to fix
the sheet.

CSV -- Excel was unable to easily open the CSV file, which was because
it was not actually comma-separated. I made it so that it is comma
separated, and now Excel can open it directly and users can cut-and
paste the data as intended.