Updated zero files (bugfix) / StudentHash comments
It turns out that smhasher provides input to StudentHash in four-byte chunks, padding odd sizes with binary zeros.

This is a problem, because if you have files that are all binary zero but increase in size by one byte each, groups of four files will be padded to be *all zeros* which will artificially make them collide. (They are not the same files, but the data that StudentHash gets is the same.) To put it another way, smhasher makes everything multiples of four bytes long, and pads with zeros.

The "right" fix to this would be to have smhasher give the original data in a uint8_t array and let the hash deal with odd sizes, but I guess the smhasher designer did not want to do that.

The fix in this commit makes all zero files increase in four-byte increments. Since there are no "odd sized" files, then there won't be any padding issue.

I also added some comments in StudentHash.